+++ src/AdvertBoard/Application/AdvertBoard.Application.AppServices/Contexts/Users/Services/UserService.cs
@@ -1,11 +1,11 @@
-﻿using System.Security.Claims;
+﻿using AdvertBoard.Application.AppServices.Authorization.Requirements;
 using AdvertBoard.Application.AppServices.Contexts.Users.Builders;
 using AdvertBoard.Application.AppServices.Contexts.Users.Repositories;
 using AdvertBoard.Contracts.Contexts.Users;
 using AdvertBoard.Contracts.Shared;
-using AdvertBoard.Domain.Contexts.Images;
 using AdvertBoard.Domain.Contexts.Users;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 
 namespace AdvertBoard.Application.AppServices.Contexts.Users.Services;
@@ -17,21 +17,24 @@ public class UserService : IUserService
     private readonly IUserSpecificationBuilder _specificationBuilder;
     private readonly IMapper _mapper;
     private readonly IHttpContextAccessor _httpContextAccessor;
+    private readonly IAuthorizationService _authorizationService;
 
     /// <summary>
-    /// Инициализирует экземпляр класса.
+    /// Инициализирует экземпляр класса <see cref="UserService"/>.
     /// </summary>
     /// <param name="userRepository">Репозиторий.</param>
     /// <param name="specificationBuilder">Строитель спецификаций.</param>
     /// <param name="mapper">Маппер.</param>
-    /// <param name="httpContextAccessor">Http контекст.</param>
+    /// <param name="httpContextAccessor">Передает HttpContext.</param>
+    /// <param name="authorizationService">Сервис для реализации requirements.</param>
     public UserService(IUserRepository userRepository, IUserSpecificationBuilder specificationBuilder, IMapper mapper,
-        IHttpContextAccessor httpContextAccessor)
+        IHttpContextAccessor httpContextAccessor, IAuthorizationService authorizationService)
     {
         _userRepository = userRepository;
         _specificationBuilder = specificationBuilder;
         _mapper = mapper;
         _httpContextAccessor = httpContextAccessor;
+        _authorizationService = authorizationService;
     }
 
     /// <inheritdoc/>
@@ -41,23 +44,30 @@ public class UserService : IUserService
     }
 
     /// <inheritdoc/>
-    public async Task<Guid> UpdateAsync(UpdateUserDto updateUserDto, CancellationToken cancellationToken)
+    public async Task<Guid> UpdateAsync(Guid userId, UpdateUserDto updateUserDto, CancellationToken cancellationToken)
     {
-        var claims = _httpContextAccessor.HttpContext.User.Claims;
-        var claimId = claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
-        var userId = Guid.Parse(claimId);
+        var existingUser = await _userRepository.GetByIdAsync(userId, cancellationToken);
+        //TODO: Исключение
+        if (existingUser is null) throw new Exception();
+        var authResult = await _authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext.User, existingUser,
+                new ResourceOwnerRequirement());
+        //TODO: исключение
+        if (!authResult.Succeeded) throw new Exception("нет доступа");
         var user = _mapper.Map<UpdateUserDto, User>(updateUserDto);
         user.Id = userId;
-        return await _userRepository.UpdateAsync(user, cancellationToken);
+        return await _userRepository.UpdateAsync(userId, user, cancellationToken);
     }
 
     /// <inheritdoc/>
-    public Task<bool> DeleteAsync(CancellationToken cancellationToken)
+    public async Task<bool> DeleteAsync(Guid userId, CancellationToken cancellationToken)
     {
-        var claims = _httpContextAccessor.HttpContext.User.Claims;
-        var claimId = claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
-        var userId = Guid.Parse(claimId);
-        return _userRepository.DeleteAsync(userId, cancellationToken);
+        var existingUser = await _userRepository.GetByIdAsync(userId, cancellationToken);
+        //TODO: Исключение
+        if (existingUser is null) throw new Exception();
+        var authResult = await _authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext.User, existingUser,
+                new ResourceOwnerRequirement());
+        if (!authResult.Succeeded) throw new Exception("нет доступа");
+        return await _userRepository.DeleteAsync(userId, cancellationToken);
     }
 
     /// <inheritdoc/>
